fix float to double @ exercicio1

+++ 3a-lista/Exercicios/src/exercicio1/CalculoArea.java
@@ -1,5 +1,5 @@
 package exercicio1;
 
 public interface CalculoArea {
-    public float area();    
+    public double area();    
 }
